+++ examples/scripts-search-plugin/src/gui.rs
@@ -704,7 +704,7 @@ impl ScriptsSearchApp {
                 .call_edit_section(|e| {
                     let created =
                         e.create_object(&effect.effect.name, e.info.layer, e.info.frame, None)?;
-                    e.focus_object(&created)?;
+                    e.focus_object(created)?;
                     anyhow::Ok(())
                 })
                 .map_err(anyhow::Error::from);
@@ -871,7 +871,7 @@ impl ScriptsSearchApp {
                 0,
             ) {
                 Ok(created) => {
-                    edit.focus_object(&created)?;
+                    edit.focus_object(created)?;
                     anyhow::Ok(())
                 }
                 Err(err) => {
@@ -914,7 +914,7 @@ impl ScriptsSearchApp {
                 e.info.frame,
                 0,
             )?;
-            e.focus_object(&created)?;
+            e.focus_object(created)?;
 
             anyhow::Ok(())
         })?